requirements.txt
fredapi>=0.5.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
hmmlearn>=0.3.0
torch>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
python-dotenv>=1.0.0
tqdm>=4.65.0
